refactor: improve error handling
Moved `build dag` error handler as the first call after finding the
configuration file, this is because that way all subsequent calls can be sure
that there is no cycle dependency that could lead to infinite loops.

Also, changed message errors and reagrouped error handlers to make them closer
to error handlers of the same type.